Moving interfaces into their own file
